backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
google-generativeai
python-docx
PyPDF2
python-pptx
python-dotenv
openpyxl
pandas
